hex config working

+++ Site.hpp
@@ -6,7 +6,7 @@
 class Site {
 public:
     int idx;    // index 0, 1, 2, ... Ns-1
-    int x, y;   // coordinates
+    double x, y;   // coordinates
     int Sz;     // Ising spins
     int new_Sz;
     int potts;
